Lower-ceremony variant of 'Dialog'.
Note that the TripleSec code has acquired a (hopefully temporary)
dependance on the 'rst.todo' package, since that's where the android
plugin puts TypedResource.  Unfortunately, there doesn't seem to be
any easy way to change this... I may want to hack the plugin itself,
at least to make this configurable.