All DB access on a separate thread.
This particular version is a sort of "good news/bad news" arrangement.
Good news:  since I already wrapped everything in a "doChange" wrapper,
I can just have that "runOnDbThread", and changes elsewhere are minimal.
Bad news:  UI updates wait for the DB, even though there's an in-core
copy which they *could* update perfectly sensibly.  May deal with that
next.

Note also the inconsistency between runOnDbThread and runOnUiThread:
the former takes an arbitrary thunk by name; the latter's arg has to
explicitly be a function or Runnable.  (Because, as noted in a previous
log entry, a version of that which took a thunk would do the wrong
thing with a Runnable.)