Seriously glitzed-up CursorAdapter.
In addition to the useual shorthands for declaration and use (type-
safed "bindItem", etc.), this also does some useful plumbing.

In particular, if you give it a ChangeNotifier[PositronicCursor] on
construction, then it will register itself for change notifications
during the lifetime of its controlling Activity, which saves you
the trouble of wiring it up yourself for this common case.