Support for "queries" in change notifications framework.
A regular change notifier watches some model state, and sends the UI
reports when it changes.  A query does the same, except that the reports
depend in part on "query parameters" that can also be changed by the UI.

So, for instance (as will show up in a few commits), a Todo List could
provide an "items" query type with a "showDone" parameter, and arrange
for "done" items to be reported or not reported to the UI depending
on whether showDone is true or false.