Get cursors out of the UI's hair.
IndexedSeq's are easier to deal with, if all you need is a wrapped-up
version of a result set.   Which is all that cursors manipulated from the
UI are *supposed* to be, with the Android 3.0 "no I/O on the UI thread"
rules.  (If your cursor isn't supposed to be doing I/O, why do you
still have to close it?)