Try to automatically map db columns to like-named java fields.
"like-named" doesn't mean identically named; we expect db fields to
have underscore_separated_names, and the corresponding fields in
Scala records to be camel-cased.  Also "_ID" maps to "id".

Mappings can be manually specified to override the automatic ones;
also, there's a "BaseRecordManager" that doesn't try the auto-mapping
voodoo, in case we're talking to a poorly-behaved ContentProvider or
something for which it goes south.  (Typically by returning an
excessive number of rows for a query with condition "2 + 2 = 5",
or just somehow gagging on it.)