Kludge to give RecordManagers access to the new IDs of inserted records.
On the plus side:  this is necessary for a working version of the Demo
App With One Of Everything (which has a RecordManager that wants to
also manage external state --- specifically, a set of ongoing proximity
alerts which are supposed to mirror the active state of a "places" table).

On the minus side, this now just breaks with any ContentQuery for which
IDs is not Long.  I can live with that --- but it breaks at runtime, and
at the very least, it would be better for it to complain at compile time.
(Or make everything generic over IdType, but that's more awkward.)