Neaten up the sample app...
... mostly by deleting references to deleted code which I'm not putting
back because it just doesn't make sense in the context of this particular
app.  (I was using it to implement a feature no one would want, because
at the time, I didn't have support for testing the API any other way.  Now,
I've got Robolectric for that, and the tests are in place.  A demo is still,
needed, but I'll be adding it to the Call Log, where it *does* make sense.)