Proof-of-concept LensFactory implementation.
At least one style of giving UI components access to setting and
tweaking "the property with name X and type Y".

Probably too slow for actual use.  At the very least, we want to
heavily memoize this.  But it at least demonstrates dancing-bear
level feasibility.  ("It's not how well the bear dances; it's that
it dances at all"...)