UiBinder handling for Views now passing all tests.
The big internal change here is properly handling the case that some view
class (say, TextView) has a handler for properties of a particular type,
but we can't find a property that matches.  In that case, we now say
that there's no binding.  (In particular, this allows the previously
documented fallback to '_.setText(_.toString)' for TextViews to actually
work!)