Tweak fixture handling in tests of variant-record manager.
In particular, try to make it a little harder to break 'em just by adding
a new fixture.  (For large test suites, it's a losing battle, which is why
Rails app test suites usually use factories these days.  But what we're
testing here is some of the machinery you'd use to implement the factory...
and the fixtures are specific to one sub-suite of tests, not the whole
megillah.)