HasMany associations with variants as targets.
Allows us to say that a School has many People, with People being
a class hierarchy including Student and Teacher.  (Code support was
already mostly there, but I needed to smooth out a few glitches.)

The API has one minor wart, which is that defaulting the name of the
foreign key doesn't work properly.  (It's not the only one where
variants are concerned; I marked this and the previously noted one
with 'XXX' comments in the spec for later review.)