Harvest content provider column names from their declared statics.
A lot of standard content providers (e.g. the call log) expose the
names of their columns as declared static fields on some class, e.g.
Calls.NUMBER and Calls.CACHED_NAME for the call log.  It's useful to
have the ORM figure out the obvious correspondances for itself, so
they don't clutter up the code.