Support mapping fields read-only or write-once-at-insert.
The intent is to support content providers that manage some data columns
internally; we can always read these stuff, but it may be an error to set
it on an existing record (or even on a new one).  Primary keys get "read
only" mapping by default, since we expect all of our data stores to assign
them on insert.

Also comes with beefed-up testing for field mapping generally.