Serializability for RecordIds and ManagedRecords.
This comes in two flavors.  There's serialization of the RecordId and
ManagedRecord objects directly, which requires custom handling to *avoid*
trying to serialize the RecordManager.  This passes tests in Robolectric,
but bits may get left behind when they're passed in a bundle for Intents.
So, to keep things simple, you can also ask a RecordId to pickle itself,
which gives you an object that requires no special handling.