More workable serialization strategy for RecordIds...
The custom readObjectStream and writeObjectStream methods were acting
flaky when RecordIds were serialized into intents, so we instead do
the moral equivalent a slightly different way:  still marking the
reference to the record manager as transient, but now reconstructing
it on first use.

This allows us to get rid of the genuinely silly 'PickledId' nonsense.