Robolectric test glue complains when given bogus paths.
This improves on the former behavior of emitting mysterious NPEs later
on which are impossible to trace to the root cause.  (The "file not
found" exceptions are caught by Robolectric itself and thrown away;
thereafter, it NPEs using data structures that weren't initialized from
the file it never read.)

We're still vulnerable to this kind of lossage if AndroidManifest.xml
is garbled, but with luck, aapt will have griped about that.