Sample contacts startup list has a UI that makes at least a little sense.
Instead of doing the ExpandableList business even for aggregated contacts
with only one RawContact, we go straight to a singleton.  For an aggregate,
we pop up a menu asking which RawContact they want to edit.  (Standard app
lets users edit both at once; it would be fairly easy to do that, too, but
I've got future plans for UI around the aggregation stuff which might not
work well with this...)