Contacts sample: Refactor generic view-wrangling code into a 'utils' trait.
This includes code to pop up "chooser" dialogs, and search for a view's
parents or children of a particular type.  It'll probably be headed into
the general library in due course, since there's nothing about it specific
to the app.

(Also sweeps in one minor bugfix, to just ignore UnknownData when editing
a contact.)