Refactoring to allow account to determine what data categories are available.
Introduces DataKindInfo and CategoryInfo to represent what kinds of data
rows are supported by a particular account type, attaches them to the
AccountInfo, and arranges for ContactEditState to query that about the
details at appropriate points.

(This API is designed to extend to enforcement of the limits on what you
can put into an Exchange account, at least as regards numbers of data rows
of particular kinds and categories.  We *are* gonna need that...)