Refactor for consistency in nomenclature.
DataKinds are now called that everywhere; the widget that groups all
data of a particular kind is now a DataKindEditor.  The "home/work/other"...
subcategories of these are now CategoryLabels, with an integer categoryTag
and a string label as the underlying representation.

Nomenclature for the latter, in particular, had been all over the place.