Prototyping new API for letting a user choose one behavior of several.
This is initially in the sample contacts 'ViewUtils', but like all that
other stuff, I'm expecting to migrate it to the core framework eventually.

We want to say something like

  dialogResultMatch( titleResource ) {
    case R.string.choice1:
      stuff...
    case R.string.choice2:
      stuff...
  }

However, doing things literally that way doesn't let the implementation
of dialogChoiceMatch get access to the choice labels.  So, instead, we
have it take a second argument list which has a list of explicit "choice"
structures:

  dialogResultMatch( R.string.sillyDialog ) (
    dialogCase( R.string.choice1 ){
      Log.d("XXX", "Choice 1")
      Toast.makeText( getContext, R.string.choice1, Toast.LENGTH_LONG ).show
    },
    dialogCase( R.string.choice2 ){
      Log.d("XXX", "Choice 2")
      Toast.makeText( getContext, R.string.choice2, Toast.LENGTH_LONG ).show
    }
  )

The nomenclature here may also be subject to... improvement.