More robust activity-result dispatch machinery.
If we ask another app for data (via StartActivityForResult), our own process
may get killed and recreated while the other activity is running.  So,
registering lambdas as handlers is dicey --- they don't ordinarily
serialize well.

So, this patch has a revised activity-result dispatch machinery which
still allows UI subcomponents to register, with a key that says which
"equivalent" component should receive the activity result in a revenant
process.