Sample contacts app:  Simultaneously edit all raw contacts of an aggregate.
We're doing this now because there are a bunch of other things that need
a model object that represents everything in the raw contact.  (Like, for
instance, "view contact".)

For now, the presentation is an in the standard Gingerbread app, which is
to say that all of the raw contacts are stacked on top of each other.
(Alternatives are easy to imagine, but this is the simplest thing to
implement...)