Allow RecordManager declarations to say fields should default ReadOnly.
Some content providers export pseudo-tables that are mostly read-only,
with a few writable columns.  To correctly map this before, you'd need
to explicitly mark all the typical columns read-only (possibly omitting
the read-write special cases) --- which is pretty much backwards from
what's most expressive in that case.

You can now put a "defaultFieldMapping( MapAs.ReadOnly )" at the top of
a record-manager constructor, to indicate that that's the default unless
otherwise specified (and in particular, for all columns mapped
automatically).