Second cut at implementation of the DataStream monad, Duration, etc.
This version takes into account that a data stream's first value may not
be available immediately on creation.  It also defines the combinators
in ways that are, at least, less obviously wrong.  (For example, the
initial value of the stream returned by "someStream.filter( pred )"
will always satisfy the predicate, even if the most recent value of the
underlying stream, at the time of creation, does *not*.)