Factor field management away from scope in BaseRecordManager
The objective here is to support multiple record managers for a given
record type, with parameters (e.g., "parent record ID") --- but isolate
all the donkey work of record management in a RecordDataWrangler, or
something like that, which builds and memoizes all of the field-mapping
data structures once.

(Ultimate cause of this:  trying to deal with a ContentProvider-ized
version of the todo-list app, in which you can't formulate a TodoItem
query without having a particular TodoList ID prespecified, due to
the hierarchical structure of the URLs.  It may be overkill, but we'll
see.  This will all be refactored further anyway; work in progress.)