At least build for ICS and Jellybean targets.
The Cursor interface added a getType method in Honeycomb and later
which my wrappers weren't wrapping.  With that present, a sample app
still loads and runs on an Eclair emulator, so putting it in appears
to be (mostly?) harmless.

ICS is now too much of the installed base to ignore, so deal...