Partially undo "top-level scope" changes; surplus to requirements.
The idea had been to have DependentRecordManagers (next commit)
special-case ID handling --- in particular, to have extra state
associated with the IDs they manage.  But the values to use for
that state are not available in all cases (in particular, not when
the ID is being fetched out of a column in a query on some other
table from the same content provider), so we punt.