Arrange for a representation for record ID columns populated by "null".
This is experimental.  It adds the concept of "null status" to a RecordId
object, adds a "nullId" method to RecordManagers which will serve up a new
"null RecordId" on demand, and arranges for "null status" to be set on
query for fields where the underlying column is a SQL null.

The code has tests and works, but the approach is somewhat troublesome,
because it doesn't generalize easily to other types of columns.  The more
idiomatic way of dealing with this situation in Scala would be to have
the fields declared with option types, set to Some(whatever) if the
underlying column was populated, and None if it was a SQL null.  However,
that's tough to do at runtime with Scala 2.9 reflection, since at runtime
(thanks to erasure) all Option types look the same.

Scala 2.10 may change this (or may offer other ways to deal with the
situation, using macros or similar black magic), which would make it
doubly awkward to keep the special case...